refactored modules, main sctipt refreshed
